Remove redundant Coalesce Mutator and rename FlipCoalesce to Coalesce
