+++ src/Mutator/ProfileList.php
@@ -156,7 +156,6 @@ final class ProfileList
         Mutator\Operator\Coalesce::class,
         Mutator\Operator\Continue_::class,
         Mutator\Operator\Finally_::class,
-        Mutator\Operator\FlipCoalesce::class,
         Mutator\Operator\Spread::class,
         Mutator\Operator\Ternary::class,
         Mutator\Operator\Throw_::class,
@@ -346,7 +345,6 @@ final class ProfileList
         'Coalesce' => Mutator\Operator\Coalesce::class,
         'Continue_' => Mutator\Operator\Continue_::class,
         'Finally_' => Mutator\Operator\Finally_::class,
-        'FlipCoalesce' => Mutator\Operator\FlipCoalesce::class,
         'Spread' => Mutator\Operator\Spread::class,
         'Ternary' => Mutator\Operator\Ternary::class,
         'Throw_' => Mutator\Operator\Throw_::class,
